Separate nonlinear test tiers and exercise two JAX versions
